--- frontend/src/types/vue-icon.d.ts
@@ -1,4 +1,0 @@
-declare module '@jamescoyle/vue-icon' {
-  const SvgIcon: any;
-  export default SvgIcon;
-}
